\name{wheat.G}
\alias{wheat.G}
\docType{data}
\title{Wheat genotypes}
\description{
A structured population of 599 wheat lines genotyped at 1279 bi-allelic markers.
}
\usage{data(wheat.G)}

\details{
Data taken from the Bayesian Linear Regression (BLR) package and recoded as \{-1,0,1\} = \{aa,Aa,AA\}.
}
\references{
Perez et al. 2010. Genomic-enabled prediction based on molecular markers and pedigree using the
Bayesian Linear Regression package in R. Plant Genome 3:106-116.
}


