\name{wheat.y}
\alias{wheat.y}
\docType{data}
\title{Wheat yields}
\description{
Average grain yield for a structured population of 599 wheat lines.
}
\usage{data(wheat.y)}

\details{
Data taken from the Bayesian Linear Regression (BLR) package and averaged over the four environments
to give a single observation per line.
}
\references{
Perez et al. 2010. Genomic-enabled prediction based on molecular markers and pedigree using the
Bayesian Linear Regression package in R. Plant Genome 3:106-116.
}